Extensive changes.
Added a function for checking MAC address equality, checking if a MAC
address is the broadcast MAC address, and getting the payload length
from a MIP packet.

In addition, changed the construct_mip_packet function to make use of
the payload length supplied as argument, instead of checking the
supplied payload for length, to be used for copying the payload to the
packet. Also fixed the way the payload length was added to the MIP
header.

Changed the send_mip_packet function to make handling of routing packets
possible.

Added a function forward_mip_packet for forwarding complete MIP packets.

Changed the recv_mip_packet function extensively, to now handle
everything when data is received over the network, and it also includes
handling of routing data. It now also makes use of the packet queueing.
It will forward packets when necessary, and directly send data to the
connected application and router when possible.

Changed the send_mip_broadcast function to no longer wait for a response
to the broaddcast.