Now freeing the packet queues on exit.
Changed modification of the epoll events to be clearer.

Checking for if the next hop is a local mip address when forwarding, and
sending it to connected application if that is the case.

Modified some error handling when receiving data over the network, in
response to the recv_mip_packet function changing.

Added some more debugging text.