Added a function for freeing data in the distance table.
Added a debugging function for printing destinations currently in the
routing table.

Now freeing the data in the distance table on exit.

Fixed updating of known neighbours.

Fixed iteration through entries in a received routing update.

Fixed updating of routing table entry based on a received, previously
known destination to not update if the update was from the current next
hop of that destination, but the cost was neither higher nor lower than
the previous cost.

Fixed the loop which checks for next hop when responding to a forwarding
request to no longer use a variable in use elsewhere to iterate.

Added debugging.