Changed debugging messages to be less intrusive.
Created a function to deal with reception of routing updates from
router.

Created function to deal with forwarding of mip packets based on forward
response from router.

Created a function to deal with reception and forwarding of messages
received from the connected application

Added a function to handle keyboard interrupts when received.

Created function to initialize the router with local mip addresses on
startup.

Added documentation for the functions.